requirements.txt
flask
pydantic
PyYAML